+++ flatlaf-core/src/test/java/com/formdev/flatlaf/TestUIDefaultsLoader.java
@@ -84,10 +84,12 @@ public class TestUIDefaultsLoader
 	void parseBorders() {
 		Insets insets = new Insets( 1,2,3,4 );
 		assertBorderEquals( new FlatEmptyBorder( insets ), "1,2,3,4" );
+		assertBorderEquals( new FlatEmptyBorder( insets ), "1,2,3,4,,," );
 		assertBorderEquals( new FlatLineBorder( insets, Color.red ), "1,2,3,4,#f00" );
 		assertBorderEquals( new FlatLineBorder( insets, Color.red, 2.5f, 0 ), "1,2,3,4,#f00,2.5" );
 		assertBorderEquals( new FlatLineBorder( insets, Color.red, 2.5f, 6 ), "1,2,3,4,#f00,2.5,6" );
 		assertBorderEquals( new FlatLineBorder( insets, Color.red, 1, 6 ), "1,2,3,4,#f00,,6" );
+		assertBorderEquals( new FlatLineBorder( insets, null, 1, 6 ), "1,2,3,4,,,6" );
 	}
 
 	private void assertBorderEquals( Border expected, String actualStyle ) {